Use exec to run the app
Without this `exec`, when the release is run in the foreground,
the shell process that was invoked stays around for the duration of the
execution. If the process that launched the script kills the shell,
the `beam.smp` process becomes orphaned. This is problematic when
integrating with e.g. Upstart.